requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
ta>=0.10.2
pydantic>=2.0.0
pyyaml>=6.0
loguru>=0.7.0
rich>=13.0.0
click>=8.1.0
matplotlib>=3.7.0
scipy>=1.11.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
asyncio-throttle>=1.0.2
